rename(component): 공통 컴포넌트 index.ts 제거 및 import 수정
