+++ src/app/(route)/write/location/_components/_internal/BottomSheet/BottomSheet.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components";
+import { Button } from "@/components/common";
 import { Radius } from "@/types";
 import { DISTANCE_OPTIONS } from "./DISTANCE_OPTIONS";
