+++ src/components/domain/index.ts
@@ -5,3 +5,4 @@ export { default as ImageViewerModal } from "./ImageViewerModal/ImageViewerModal
 export { default as CommentList } from "./CommentList/CommentList";
 export { default as AuthLogoLink } from "./AuthLogoLink/AuthLogoLink";
 export { default as FindPwForm } from "./FindPwForm/FindPwForm";
+export { default as FooterButton } from "./FooterButton/FooterButton";
