--- src/components/index.ts
@@ -1,35 +1,0 @@
-export { default as Icon } from "./common/Icon/Icon";
-export { default as ConfirmModal } from "./common/Modal/ConfirmModal";
-export { default as Dropdown } from "./common/Dropdown/Dropdown";
-export { default as Toast } from "./common/Toast/Toast";
-export { default as Input } from "./common/Input/InputText/InputText";
-export { default as Chip } from "./common/Chip/Chip";
-export { default as CheckBox } from "./common/CheckBox/CheckBox";
-export { default as Badge } from "./common/Badge/Badge";
-export { default as Button } from "./common/Buttons/Button/Button";
-export { default as Bookmark } from "./common/Buttons/Bookmark/Bookmark";
-export { default as KebabMenuButton } from "./common/Buttons/KebabMenuButton/KebabMenuButton";
-export { default as Location } from "./common/Buttons/Location/Location";
-export { default as ViewMoreReply } from "./common/Buttons/ViewMoreReply/ViewMoreReply";
-export { default as ViewMoreComment } from "./common/Buttons/ViewMoreComment/ViewMoreComment";
-export { default as Filter } from "./common/Buttons/Filter/Filter";
-export { default as KebabMenu } from "./common/Buttons/KebabMenu/KebabMenu";
-export { default as FloatingButton } from "./common/Buttons/FloatingButton/FloatingButton";
-export { default as ToggleButton } from "./common/Buttons/ToggleButton/ToggleButton";
-export { default as ImageSelectButton } from "./common/Buttons/ImageSelectButton/ImageSelectButton";
-export { default as InputText } from "./common/Input/InputText/InputText";
-export { default as InputField } from "./common/Input/InputField/InputField";
-export { default as InputChat } from "./common/Input/InputChat/InputChat";
-export { default as InputSearch } from "./common/Input/InputSearch/InputSearch";
-export { default as RequiredText } from "./common/RequiredText/RequiredText";
-export { default as Tab } from "./domain/Tab/Tab";
-export { default as ListSearch } from "./domain/ListSearch/ListSearch";
-export { default as PopupLayout } from "./domain/PopupLayout/PopupLayout";
-export { default as ImageViewerModal } from "./domain/ImageViewerModal/ImageViewerModal";
-export { default as CommentList } from "./domain/CommentList/CommentList";
-export { default as Header } from "./layout/Header/Header";
-export { default as Footer } from "./layout/Footer/Footer";
-export { default as DetailHeader } from "./layout/DetailHeader/DetailHeader";
-export { default as AuthLogoLink } from "./domain/AuthLogoLink/AuthLogoLink";
-export { default as FindPwForm } from "./domain/FindPwForm/FindPwForm";
-export { default as FooterButton } from "./domain/FooterButton/FooterButton";
